+++ src/docs/pages/guide/logs/index.vue
@@ -18,6 +18,37 @@
     
     ### 更新记录
 
+    ### 0.11.4 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
+
+    <ui-textcolor color="light-blue">
+        <ui-small>
+            发布于2018年7月1日，
+            Archive :
+            <a href="https://github.com/Morning-UI/morning-ui/releases/tag/0.11.4" target="_blank">Package</a>
+            <a href="https://cdn.jsdelivr.net/npm/morning-ui@0.11.4/dist/" target="_blank">jsdelivr(CDN)</a>
+            <a href="https://unpkg.com/morning-ui@0.11.4/dist/" target="_blank">unpkg(CDN)</a>
+        </ui-small>
+    </ui-textcolor>
+
+    #### 新特性
+
+    - 新增`ui-collapse`折叠面板组件
+    - `ui-upload`支持显示真实的上传进度
+    - `ui-colorpicker`新增`show-picker`、`hide-picker`、`hue-slider-change`、`alpha-slider-change`、`input-type-change`事件和`togglePicker`方法
+    - `ui-tab`新增`disabled-options`配置，支持禁用指定标签
+
+    #### 改进
+
+    - 使用Parallelism优化CircleCI持续集成性能
+
+    #### 修复
+
+    - 修复`ui-btn`和`ui-link`的一些样式问题
+    - 修复`ui-datepicker`和`ui-datetimepicker`在选中特定日期范围时显示区间错误的问题
+    - 修复`ui-dropdown`点击触发元素无法关闭下拉菜单的问题
+
+    ---
+
     ### 0.11.3 <ui-badge class="circle ver-canary" size="s" color="light-theme">Canary</ui-badge>
 
     <ui-textcolor color="light-blue">