+++ src/lib/components/tab/index.vue
@@ -8,6 +8,7 @@
         :append="append"
         :anchor-target="anchorTarget"
         :position="position"
+        :disabled-options="disabledOptions"
     >
 
         <ul>
@@ -15,6 +16,9 @@
                 v-for="item in data.namelist"
                 :name="item.name"
                 :key="item.name"
+                :class="{
+                    disabled : (conf.disabledOptions.indexOf(item.name) !== -1)
+                }"
                 v-html="item.html"
                 @click="_onClick(item.name)"
             ></li>
@@ -66,6 +70,10 @@ export default {
                 'top',
                 'left'
             ].indexOf(value) !== -1)
+        },
+        disabledOptions : {
+            type : Array,
+            default : (() => [])
         }
     },
     computed : {
@@ -76,7 +84,8 @@ export default {
                 prepend : this.prepend,
                 append : this.append,
                 anchorTarget : this.anchorTarget,
-                position : this.position
+                position : this.position,
+                disabledOptions : this.disabledOptions
             };
 
         },
@@ -104,6 +113,12 @@ export default {
     },
     methods : {
         _onClick : function (name) {
+
+            if (this.conf.disabledOptions.indexOf(name) !== -1) {
+
+                return;
+
+            }
 
             this.switch(name);
 
@@ -155,6 +170,27 @@ export default {
             this._initSelectTab();
 
         },
+        _getClosestCanSelectTab : function (name) {
+
+            if (this.conf.disabledOptions.indexOf(name) !== -1) {
+
+                if (this.data.tabs.indexOf(name) < (this.data.tabs.length - 1)) {
+
+                    return this.data.tabs[this.data.tabs.indexOf(name) + 1];
+
+                } else if (this.data.tabs.indexOf(name) === (this.data.tabs.length - 1)) {
+
+                    return this.data.tabs[0];
+
+                }
+
+                return null;
+
+            }
+
+            return name;
+
+        },
         _initSelectTab : function () {
 
             this.Vue.nextTick(() => {
@@ -189,11 +225,11 @@ export default {
                 // init new select tab
                 if (this.conf.tab) {
 
-                    this.switch(this.conf.tab);
+                    this.switch(this._getClosestCanSelectTab(this.conf.tab));
 
                 } else {
 
-                    this.switch(this.data.tabs[0]);
+                    this.switch(this._getClosestCanSelectTab(this.data.tabs[0]));
 
                 }
 
@@ -403,7 +439,7 @@ export default {
 
         });
 
-        this.$watch('conf.tab', () => {
+        this.$watch(() => (`${JSON.stringify(this.conf.disabledOptions)}||${this.conf.tab}`), () => {
 
             this._initSelectTab();