+++ src/lib/components/upload/index.vue
@@ -49,7 +49,13 @@
                     :key="index"
                     :class="item.classList"
                 >
-                    <i class="progress"></i>
+                    <i
+                        class="progress"
+                        :class="item.classList"
+                        :style="{
+                            width : (item.classList.uploading) ? ((30 + (+item.progress) * 60) + '%') : 'auto'
+                        }"
+                    ></i>
                     <span>
                         {{item.name}}
                     </span>
@@ -414,6 +420,7 @@ export default {
                 size : 0,
                 data : undefined,
                 path : undefined,
+                progress : 0,
                 classList : {
                     fail : false,
                     uploading : false,
@@ -544,7 +551,25 @@ export default {
             }
 
             let index = this.data.uploadQueue.shift(),
-                uploadObj = {};
+                uploadObj = {
+                    onUploadProgress : xhr => {
+
+                        if (xhr &&
+                            xhr.upload &&
+                            typeof xhr.upload.addEventListener === 'function') {
+
+                            xhr.upload.addEventListener('progress', evt => {
+
+                                this.data.files[index].progress = +(evt.loaded / evt.total) || 0;
+
+                            }, false);
+
+                        }
+
+                    },
+                    file : null,
+                    name : null
+                };
 
             Promise
                 .resolve()